feat: Removing helper function. (#61)
* Addressing @wfahim comments in separate PR

* feat: Removing helper function Update Plotly code for v1.0 #60

* converting to px instead of points

* adding references

* updating coockbook to remove titles and use default pio template

* updating tests and  readme

* adding #| output: false

* removing axis title

* removing theme_af.py
