+++ .github/workflows/deploy.yml
@@ -5,7 +5,7 @@ on:
   
   workflow_dispatch:
   push:
-    branches: [main, dev]
+    branches: [main, dev, fr/plotly_review]
 
 jobs:
