+++ README.md
@@ -22,11 +22,50 @@ pip install afcharts-py
 
 ## Usage
 
+### Matplotlib
+
 To use the afcharts style in matplotlib, import the package, then simple use the afcharts style with
+
 ```python
 import afcharts
 import matplotlib.pytplot as plt
 plt.style.use('afcharts')
+```
+
+### Plotly
+
+Basic example of using the plotly pio template to produce charts.
+
+```python
+
+import pandas as pd
+import plotly.express as px
+import plotly.graph_objects as go
+
+# AF Package
+from afcharts.pio_template import pio
+
+# Set default theme
+pio.templates.default = "theme_af"
+
+
+# Load the gapminder dataset from plotly.express
+df = px.data.gapminder().query("year == 2007 & continent == 'Americas'")
+
+top5 = df.nlargest(5, "pop")
+
+fig = go.Figure()
+
+fig.add_trace(
+    go.Bar(
+        x=top5["country"],
+        y=top5["pop"],  # Repeat the category name for each x value
+        name="test bar",  # Use the x value column name as the trace name
+    )
+)
+
+fig.show()
+
 ```
 
 then create your chart as normal.
