+++ src/afcharts/cookbook/03-plotly-usage.qmd
@@ -1,72 +1,84 @@
 # Using the package with Plotly
-## Example of a bar chart
+## Bar charts
 
 ```{python}
-from afcharts.theme_af import theme_af
-import plotly.graph_objects as go
+#| eval: true
+#| output: false
+
 import pandas as pd
 import plotly.express as px
+import plotly.graph_objects as go
+
+# AF Package
+from afcharts.pio_template import pio
+
+# Set default theme
+pio.templates.default = "theme_af"
 
 # Load the gapminder dataset from plotly.express
 df = px.data.gapminder().query("year == 2007 & continent == 'Americas'")
 
-
 top5 = df.nlargest(5, "pop")
-
 fig = go.Figure()
 
 fig.add_trace(
     go.Bar(
         x=top5["country"],
         y=top5["pop"],  # Repeat the category name for each x value
-        name="test bar",  # Use the x value column name as the trace name
+
     )
 )
 
 # Update layout
 fig.update_layout(
-    template=theme_af(grid="y"),
-    title=dict(
-        text="The U.S.A. is the most populous country in\nthe Americas",
-        subtitle=dict(text="Population of countries in the Americas, 2007"),
-    ),
-    margin=dict(b=70),  # Increase bottom margin to make space for source
-    annotations=[
-        dict(
-            text="Source: Gapminder",
-            xref="paper",
-            yref="paper",
-            x=0,
-            y=-0.1,
-            xanchor="left",
-        )
-    ],
-    meta=dict(alt="Bar chart showing sales of Product A (23 units), Product B (45 units), and Product C (56 units)."),
+    height=300
 )
 
 fig.show()
+
 ```
+
+<div class="chart-title">The U.S.A. is the most populous country in the Americas</div>
+<div class="chart-subtitle">Population of countries in the Americas (millions), 2007</div>
+
+
+```{python}
+#| echo: false
+
+fig.show()
+
+```
+
+<div class="chart-source">Source: Gapminder</div>
+
+This bar chart uses the afcharts theme, and shows the populations of the five most populous countries in the Americas. Each bar is dark blue and labelled by country underneath. All text is black in a sans serif font. Pale grey grid lines extend out from the y axis.
 
 ## Example of a line chart with duo palette
 
 ```{python}
-import plotly.graph_objects as go
+#| label: line-charts-2
+#| eval: true
+#| output: false
+
 import pandas as pd
 import plotly.express as px
+import plotly.graph_objects as go
 
 # AF Package
-from afcharts.theme_af import theme_af
+from afcharts.pio_template import pio
 from afcharts.assets.af_colours import duo
+
+# Set default theme
+pio.templates.default = "theme_af"
 
 # Load the gapminder dataset from plotly.express
 df = px.data.gapminder()
 
 df = df[df["country"].isin(["United Kingdom", "China"])]
-
 fig = go.Figure()
 
 # Add a trace for each continent
-for country in df["country"].unique():
+for i, country in enumerate(df["country"].unique()):
     df_country = df[df["country"] == country]
     fig.add_trace(
         go.Scatter(
@@ -75,48 +87,66 @@ for country in df["country"].unique():
             mode="lines",
             name=country,
             text=df_country["country"],
+            line=dict(color=duo[i % len(duo)])
         )
     )
 
 # Update layout
 fig.update_layout(
-    template=theme_af(grid="y", colour_palette=duo),
     xaxis=dict(
-        title="Year",
+        showgrid=False  # Hide x-axis grid lines
     ),
-    title=dict(
-        text="Living Longer",
-        subtitle=dict(text="Life Expectancy in the United Kingdom and China 1952-2007"),
+    yaxis=dict(
+        range=[0, 82],
+        tickmode="linear",
+        dtick=20,  # Show ticks every 20 units
     ),
-    margin=dict(b=70),  # Increase bottom margin to make space for source
-    annotations=[
-        dict(
-            text="Source: Gapminder",
-            xref="paper",
-            yref="paper",
-            x=0,
-            y=-0.1,
-            xanchor="left",
-        )
-    ],
-    meta=dict(alt="Bar chart showing sales of Product A (23 units), Product B (45 units), and Product C (56 units)."),
+    legend=dict(
+        orientation="h",
+        yanchor="top",
+        y=-0.2,
+        xanchor="center",
+        x=0.5
+    ),
+    height=300
 )
 
-fig.update_yaxes(range=[0, 82])
-
 fig.show()
+
 ```
 
-## Example of a scatter chart
+<div class="chart-title">Living Longer</div>
+<div class="chart-subtitle">Life Expectancy in the United Kingdom and China 1952-2007</div>
 
 ```{python}
-import plotly.graph_objects as go
+#| echo: false
+
+fig.show()
+
+```
+
+<div class="chart-source">Source: Gapminder</div>
+
+
+This line chart uses the afcharts theme and there are thin pale grey lines extending from the y axis. There are two thicker lines showing the life expectancy in the UK and China over time. The line colours are from the main Analysis Function palette - dark blue for China and orange for the UK, denoted by a legend at the bottom of the chart.
+
+Legends should be avoided unless absolutely necessary, as these usually rely on using colour to match labels to data. More information can be found in the Analysis Function [charts guidance](https://analysisfunction.civilservice.gov.uk/policy-store/data-visualisation-charts/#section-11). It is best practice to label lines directly, and an example of this can be found in the [annotations](#annotations) section.
+
+## Example of a scatterplot
+
+```{python}
+#| eval: true
+#| output: false
+
 import pandas as pd
 import plotly.express as px
+import plotly.graph_objects as go
 
 # AF Package
-from afcharts.theme_af import theme_af
-from afcharts.assets.af_colours import duo
+import afcharts.pio_template
+
+# Set default theme
+pio.templates.default = "theme_af"
 
 # Load the gapminder dataset from plotly.express
 df = px.data.gapminder()
@@ -135,28 +165,26 @@ fig.add_trace(
 
 # Update layout
 fig.update_layout(
-    template=theme_af(grid="xy", colour_palette=duo),
     xaxis=dict(
         title="GDP (US$, inflation-adjusted)",
     ),
-    yaxis=dict(
-        title="Life\nExpectancy",
-    ),
-    title=dict(
-        text="The relationship between GDP and Life Expectancy is complex",
-        subtitle=dict(text="GDP and Life Expectancy for all countires, 2007"),
-    ),
-    annotations=[
-        dict(
-            text="Source: Gapminder",
-            xref="paper",
-            yref="paper",
-            x=0,
-            y=-0.1,
-            xanchor="left",
-        )
-    ],
+    height=300
 )
 
 fig.show()
+
 ```
+
+<div class="chart-title">The relationship between GDP and Life Expectancy is complex</div>
+<div class="chart-subtitle">GDP and Life Expectancy for all countries, 2007</div>
+
+```{python}
+#| echo: false
+
+fig.show()
+
+```
+
+<div class="chart-source">Source: Gapminder</div>
+
+This scatterplot uses the afcharts theme, and shows life expectancy against GDP for all countries. Thin pale grey lines extend out from the x and y axis labels, forming a grid. The data points are plotted as dark blue circles. Both axes are labeled in black using a sans serif font.
