+++ src/afcharts/cookbook/styles.css
@@ -1,0 +1,12 @@
+.chart-source {
+    font-size: 12pt;
+}
+
+.chart-subtitle {
+    font-size: 14pt;
+}
+
+.chart-title {
+    font-size: 16pt;
+    font-weight: 700;
+}
