+++ tests/plotly_line_test.py
@@ -2,15 +2,16 @@ import os
 
 import plotly.express as px
 import plotly.graph_objects as go
-import plotly.io as pio
-
-from afcharts.assets.af_colours import duo
 
 # AF Package
-from afcharts.theme_af import theme_af
+from afcharts.assets.af_colours import duo
+from afcharts.pio_template import pio
+
+# Set default theme
+pio.templates.default = "theme_af"
 
 
-def test_plolty_line_chart():
+def test_plotly_line_chart():
     # Only set the renderer to "browser" if not running in CI
     if not os.environ.get("CI"):
         pio.renderers.default = "browser"
@@ -23,7 +24,7 @@ def test_plolty_line_chart():
     fig = go.Figure()
 
     # Add a trace for each continent
-    for country in df["country"].unique():
+    for i, country in enumerate(df["country"].unique()):
         df_country = df[df["country"] == country]
         fig.add_trace(
             go.Scatter(
@@ -32,36 +33,23 @@ def test_plolty_line_chart():
                 mode="lines",
                 name=country,
                 text=df_country["country"],
+                line=dict(color=duo[i % len(duo)]),
             )
         )
 
     # Update layout
     fig.update_layout(
-        template=theme_af(grid="y", colour_palette=duo),
         xaxis=dict(
-            title="Year",
+            showgrid=False,  # Hide x-axis grid lines
         ),
-        title=dict(
-            text="Living Longer",
-            subtitle=dict(text="Life Expectancy in the United Kingdom and China 1952-2007"),
+        yaxis=dict(
+            title="Life Expectancy",
+            range=[0, 82],
+            tickmode="linear",
+            dtick=20,  # Show ticks every 20 units
         ),
-        margin=dict(b=70),  # Increase bottom margin to make space for source
-        annotations=[
-            dict(
-                text="Source: Gapminder",
-                xref="paper",
-                yref="paper",
-                x=0,
-                y=-0.1,
-                xanchor="left",
-            )
-        ],
-        meta=dict(
-            alt="Bar chart showing sales of Product A (23 units), Product B (45 units), and Product C (56 units)."
-        ),
+        height=300,
     )
-
-    fig.update_yaxes(range=[0, 82])
 
     if not os.environ.get("CI"):
         fig.show()
