+++ tests/plotly_scatter_test.py
@@ -2,15 +2,15 @@ import os
 
 import plotly.express as px
 import plotly.graph_objects as go
-import plotly.io as pio
-
-from afcharts.assets.af_colours import duo
 
 # AF Package
-from afcharts.theme_af import theme_af
+from afcharts.pio_template import pio
+
+# Set default theme
+pio.templates.default = "theme_af"
 
 
-def test_plotly_line_chart():
+def test_plotly_scatter_chart():
     # Only set the renderer to "browser" if not running in CI
     if not os.environ.get("CI"):
         pio.renderers.default = "browser"
@@ -32,27 +32,13 @@ def test_plotly_line_chart():
 
     # Update layout
     fig.update_layout(
-        template=theme_af(grid="xy", colour_palette=duo),
         xaxis=dict(
             title="GDP (US$, inflation-adjusted)",
         ),
         yaxis=dict(
             title="Life\nExpectancy",
         ),
-        title=dict(
-            text="The relationship between GDP and Life Expectancy is complex",
-            subtitle=dict(text="GDP and Life Expectancy for all countires, 2007"),
-        ),
-        annotations=[
-            dict(
-                text="Source: Gapminder",
-                xref="paper",
-                yref="paper",
-                x=0,
-                y=-0.1,
-                xanchor="left",
-            )
-        ],
+        height=300,
     )
 
     if not os.environ.get("CI"):
